requirements.txt
Django
djangorestframework
googlemaps
psycopg2